Organizacion de las rutas y  de los respectivos endpoins
