--- server/src/routes/properties/propertiesRoutes.js
@@ -1,24 +1,0 @@
-import express from 'express';
-import propertyDetailsController from '../../controllers/owner/propertyDetailsController.js';
-import propertyExistsMiddleware from '../../middlewares/propertyExistsMiddleware.js';
-import authUserMiddleware from '../../middlewares/authUserMiddleware.js';
-import propertyController from '../../controllers/properties/propertyController.js';
-
-const router = express.Router();
-
-// Listado de propiedades
-router.get('/properties');
-// Creación de nueva propiedad
-router.post('/properties', authUserMiddleware, propertyController);
-// Detalle de una propiedad
-router.get(
-	'/properties/:propertyId',
-	propertyExistsMiddleware,
-	propertyDetailsController
-);
-// Cambio de estado de propiedad (disponible / no disponible)
-router.patch('/properties/:id');
-// Modificar una propiedad (dueño o admin) [EXTRA]
-router.put('/properties/:id');
-
-export default router;
