+++ server/src/routes/propertiesRoutes.js
@@ -4,30 +4,40 @@ import express from 'express';
 // middlewares
 import propertyExistsMiddleware from '../../middlewares/propertyExistsMiddleware.js';
 import authUserMiddleware from '../../middlewares/authUserMiddleware.js';
+import { fileUploadMiddleware } from '../middlewares/fileUploadMiddleware.js';
+import authOwnerMiddleware from '../../middlewares/authOwnerMiddleware.js';
 
 // controladores
 import propertyDetailsController from '../../controllers/owner/propertyDetailsController.js';
 import propertyController from '../../controllers/properties/propertyController.js';
+import fileUploadController from '../controllers/owner/fileUploadController.js';
 
 const router = express.Router();
 
-// Listado de propiedades
+// 09 Listado de propiedades ⛔
 router.get('/properties');
 
-// Creación de nueva propiedad ✅
+// 10 Creación de nueva propiedad ✅
 router.post('/properties', authUserMiddleware, propertyController);
 
-// Detalle de una propiedad
+// 11 Detalle de una propiedad ✅
 router.get(
-	'/properties/:propertyId',
+	'/properties/:id',
 	propertyExistsMiddleware,
 	propertyDetailsController
 );
 
-// Cambio de estado de propiedad (disponible / no disponible)
-router.patch('/properties/:id');
+// 12 Cambio de estado de propiedad (disponible / no disponible) ✅
+router.patch('/properties/:id', authUserMiddleware, propertyController);
 
-// Modificar una propiedad (dueño o admin) [EXTRA]
-router.put('/properties/:id');
+// 13 Modificar una propiedad (dueño o admin) [EXTRA] ⛔
+router.put('/properties/:id', authOwnerMiddleware, propertyController);
+
+// 14 Ruta para subir imágenes y videos asociados a una propiedad
+router.post(
+	'/upload-files/:propertyId',
+	fileUploadMiddleware,
+	fileUploadController
+);
 
 export default router;
